rename genai.* to openvino_pipeline.*
